Change urdf_parser_py import to upstream version
The local version was added in 2015 because the upstream urdf parser
didn't handle hardwareInterfaces in the transmission joints and actuators.

In the current version, the upstream urdf parser handles TransmissionJoints with
optional hardwareInterface elements.

Hardware interfaces in the actuator element shouldn't be used anymore and
are not used in the shadow repos.