provide ompl configuration for left/right_hand
Previously there was simply no configuration for
our group "right_hand" and adding one to the template
didn't work because of the assumption that there is a prefix for all groups.

I removed the assumption and provide a useful configuration
"hand" in the template that maps to left_hand/right_hand as required.

The important parameter here is again longest_valid_segment_fraction.
OMPL's default here is 0.01, which is awful for this group and results in
index/middle finger to move right through the thumb in planned trajectories.
0.001 is a reasonable value applying a 5° maximum L0 distance.